# openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A simple pastebin service with analytics
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Pastebin API
  version: "1.0"
paths:
...
  /profile:
    get:
      description: Retrieve the profile of the authenticated user
      responses:
        "200":
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.User"
        "500":
          description: Failed to get profile
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Get user profile
      tags:
        - profile
    put:
      description: Update the profile of the authenticated user
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/models.PatchProfile"
        description: Profile update data
        required: true
      responses:
        "200":
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.User"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Failed to update profile
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Update user profile
      tags:
        - profile
...
servers:
  - url: //localhost:8080
components:
  securitySchemes:
    BearerAuth:
      description: Type "Bearer" followed by a space and JWT token.
      in: header
      name: Authorization
      type: apiKey
  schemas:
...
    models.PatchProfile:
      properties:
        avatar:
          type: string
        name:
          type: string
      type: object
...
    models.User:
      properties:
        avatar:
          type: string
        email:
          type: string
        id:
          type: string
        is_verified:
          type: boolean
        name:
          type: string
      type: object